refactor: move IsManagedDomain function to dns package
